pyproject.toml
[tool.poetry]
name = "image-to-string"
version = "0.1.0"
description = ""
authors = ["chihwei"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
setuptools = "^68.2.2"
wheel = "^0.41.2"
pytesseract = "^0.3.10"
pillow = "^10.1.0"
click = "^8.1.7"
fastapi = "^0.104.0"
python-multipart = "^0.0.6"
uvicorn = "^0.23.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
